Introduce MPI ops in a new contrib project.
This commit primarily affects the build system, which now builds a new
contrib project called 'mpi'. Right now, this requires the user to set
up a third_party/mpi directory with symlinks to mpi.h,
mpi_platform_independent.h, and libmpi.dylib; in a future commit these
symlinks will be established by configure.

The ops introduced in this commit are MPISize and MPIRank, which return
int32 scalars of the global MPI size and the calling process MPI rank
and take no arguments.